Rename `doc_comment_double_space_linebreak` -> `doc_comment_double_space_linebreaks`
