+++ tests/ui/doc/doc_comment_double_space_linebreaks.stderr
@@ -1,15 +1,15 @@
 error: doc comment uses two spaces for a hard line break
-  --> tests/ui/doc/doc_comment_double_space_linebreak.rs:7:43
+  --> tests/ui/doc/doc_comment_double_space_linebreaks.rs:7:43
    |
 LL | //! Should warn on double space linebreaks  
    |                                           ^^
    |
    = help: replace this double space with a backslash: `\`
    = note: `-D clippy::doc-comment-double-space-linebreak` implied by `-D warnings`
-   = help: to override `-D warnings` add `#[allow(clippy::doc_comment_double_space_linebreak)]`
+   = help: to override `-D warnings` add `#[allow(clippy::doc_comment_double_space_linebreaks)]`
 
 error: doc comment uses two spaces for a hard line break
-  --> tests/ui/doc/doc_comment_double_space_linebreak.rs:35:51
+  --> tests/ui/doc/doc_comment_double_space_linebreaks.rs:35:51
    |
 LL | /// Should warn when doc comment uses double space  
    |                                                   ^^
@@ -19,7 +19,7 @@ LL | /// as a line-break, even when there are multiple
    = help: replace this double space with a backslash: `\`
 
 error: doc comment uses two spaces for a hard line break
-  --> tests/ui/doc/doc_comment_double_space_linebreak.rs:44:12
+  --> tests/ui/doc/doc_comment_double_space_linebreaks.rs:44:12
    |
 LL | /// 🌹 are 🟥  
    |              ^^
@@ -33,7 +33,7 @@ LL | /// and so are 🫵
    = help: replace this double space with a backslash: `\`
 
 error: doc comment uses two spaces for a hard line break
-  --> tests/ui/doc/doc_comment_double_space_linebreak.rs:86:16
+  --> tests/ui/doc/doc_comment_double_space_linebreaks.rs:86:16
    |
 LL | /// here we mix  
    |                ^^
